Refactor and add to zarr loader.
