+++ src/loaders/utils.js
@@ -53,8 +53,24 @@ export function padTileWithZeros(tile, targetWidth, targetHeight) {
  * @param {Object} tile { x, y, z }
  * @returns {TypedArray} TypedArray
  */
-export function truncateTiles(data, height, width, tileSize) {
-  return data.map((d) => {
+export function truncateTiles(data, loader, tile) {
+  const { x, y, z } = tile
+  const { tileSize } = loader;
+   let height = tileSize;
+   let width = tileSize;
+   const size = loader.getRasterSize({ z });
+   const numTilesX = Math.ceil(size.width / tileSize);
+   const numTilesY = Math.ceil(size.height / tileSize);
+   if (x === numTilesX - 1) {
+     const paddedWidth = numTilesX * tileSize;
+     width = tileSize - (paddedWidth - size.width);
+   }
+   if (y === numTilesY - 1) {
+     const paddedHeight = numTilesY * tileSize;
+     height = tileSize - (paddedHeight - size.height);
+   }
+  const tileData = { height, width }
+  tileData.data = data.map((d) => {
     let truncated = d;
     if ((width < tileSize || height < tileSize) && d.length !== width * height) {
       truncated = new d.constructor(height * width);
@@ -66,6 +82,7 @@ export function truncateTiles(data, height, width, tileSize) {
     }
     return truncated;
   })
+  return tileData;
 }
 
 /**
